timer: Explicitly cast timer value to 64-bit
The complete timer value is split between a 32-bit lower and upper half, which
means reading it should return a 64bit value overall. Cast the last
mmio_read() to comply with the compiler's requirements.

+++ timer.c
@@ -19,7 +19,7 @@ static unsigned long read_timer64(struct timer *t)
 	unsigned long val = 0;
 
 	val = mmio_read(t->lsb);
-	val |= (mmio_read(t->msb) << 32);
+	val |= ((u64)mmio_read(t->msb) << 32);
 
 	return val;
 }
